+++ mdx-components.ts
@@ -1,7 +1,0 @@
-import type { MDXComponents } from 'mdx/types';
-
-export const useMDXComponents: (components: MDXComponents) => MDXComponents = (components) => {
-  return {
-    ...components,
-  };
-};