+++ next.config.mjs
@@ -1,12 +1,12 @@
 import mdx from '@next/mdx';
 import slug from 'rehype-slug';
+import tableOfContents from 'remark-toc';
 import TerserPlugin from 'terser-webpack-plugin';
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   experimental: {
     appDir: true,
-    scrollRestoration: true,
   },
   pageExtensions: ['mdx', 'tsx'],
   reactStrictMode: true,
@@ -158,6 +158,7 @@ const withMDX = mdx({
   extension: /\.mdx?$/,
   options: {
     rehypePlugins: [slug],
+    remarkPlugins: [tableOfContents],
   },
 });