+++ src/components/Dropdown/DropdownItem.tsx
@@ -1,18 +1,21 @@
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-
+import { useListItem } from '@floating-ui/react';
 import type { ComponentProps, FC, PropsWithChildren } from 'react';
+import { useContext } from 'react';
 import { twMerge } from 'tailwind-merge';
 import type { DeepPartial } from '../../';
 import { useTheme } from '../../';
 import { mergeDeep } from '../../helpers/merge-deep';
+import type { ButtonBaseProps } from '../Button/ButtonBase';
+import { ButtonBase } from '../Button/ButtonBase';
+import { DropdownContext } from './Dropdown';
 
 export interface FlowbiteDropdownItemTheme {
+  container: string;
   base: string;
   icon: string;
 }
 
-export interface DropdownItemProps extends PropsWithChildren, ComponentProps<'li'> {
+export interface DropdownItemProps extends PropsWithChildren, ButtonBaseProps, Record<string, unknown> {
   icon?: FC<ComponentProps<'svg'>>;
   onClick?: () => void;
   theme?: DeepPartial<FlowbiteDropdownItemTheme>;
@@ -26,12 +29,29 @@ export const DropdownItem: FC<DropdownItemProps> = ({
   theme: customTheme = {},
   ...props
 }) => {
+  const { ref, index } = useListItem({ label: typeof children === 'string' ? children : undefined });
+  const { activeIndex, dismissOnClick, getItemProps, handleSelect } = useContext(DropdownContext);
+  const isActive = activeIndex === index;
+
   const theme = mergeDeep(useTheme().theme.dropdown.floating.item, customTheme);
 
   return (
-    <li className={twMerge(theme.base, className)} onClick={onClick} {...props}>
-      {Icon && <Icon className={theme.icon} />}
-      {children}
+    <li role="menuitem" className={theme.container}>
+      <ButtonBase
+        ref={ref}
+        className={twMerge(theme.base, className)}
+        {...props}
+        {...getItemProps({
+          onClick: () => {
+            onClick && onClick();
+            dismissOnClick && handleSelect(null);
+          },
+        })}
+        tabIndex={isActive ? 0 : -1}
+      >
+        {Icon && <Icon className={theme.icon} />}
+        {children}
+      </ButtonBase>
     </li>
   );
 };